// types/eventsList.ts
export enum Reports {
    GetAll = "reports/get/all",
    GetChunk = "reports/get/chunk",
    GetCount = "reports/get/count",
    GetToday = "reports/get/today",
    GetDay = "reports/get/day",
    GetPeriod = "reports/get/period",
    Delete = "reports/delete"
}

export enum Report {
    Add = "report/add",
    Get = "report/get",
    Edit = "report/edit",
    Delete = "report/delete"
}

export enum Clipboard {
    Toggle = "clipboard/toggle",
    SetState = "clipboard/setState"
}

export enum Utility {
    GetPath = "Utility/path/get"
}